carmel 6.1
git-svn-id: https://nlg0.isi.edu/svn/sbmt/trunk/graehl@2829 7f356e55-ce0a-0410-be72-bc6dfe96aa1c

+++ carmel/src/cached_derivs.h
@@ -27,7 +27,7 @@ struct cached_derivs
     {
         return cached?derivs.size():corpus.size();
     }
-    unsigned n_output() const
+    double n_output() const
     {
         return corpus.n_output;
     }
